Finding the follow user bug and fix
Logic error while following a user:
It allows to follow the same user multiple times, instead of the
proposed toggle to follow and unfollow.
Fixed: incorrect use of Key in
getFollowingUsersForUser(user:completionBlock:) in ParseHelper.

Instead of getting the following base for this user, it queried the
following base for the toUser in the Follow object, which allowed to
follow itself.